requirements.txt
ansicon==1.89.0
arrow==1.3.0
asgiref==3.8.1
astroid==3.3.8
attrs==24.3.0
audioread==3.0.1
black==24.10.0
blessed==1.20.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudinary==1.41.0
colorama==0.4.6
decorator==5.1.1
dill==0.3.9
dj-database-url==2.3.0
Django==5.1.3
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-modeltranslation==0.19.11
django-picklefield==3.2
django-q2==1.7.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
flake8==7.1.1
gTTS==2.5.4
gunicorn==23.0.0
h11==0.14.0
idna==3.10
inflection==0.5.1
isort==5.13.2
jinxed==1.3.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
mccabe==0.7.0
msgpack==1.1.0
mutagen==1.47.0
mypy-extensions==1.0.0
numba==0.60.0
numpy==2.0.2
packaging==24.2
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
postgres==4.0
psycopg2-binary==2.9.10
psycopg2-pool==1.2
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.10.0
pylint==3.3.3
pyTelegramBotAPI==4.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==3.5.3
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
scikit-learn==1.6.0
scipy==1.14.1
setuptools==75.6.0
six==1.16.0
soundfile==0.12.1
soxr==0.5.0.post1
sqlparse==0.5.2
threadpoolctl==3.5.0
tomlkit==0.13.2
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.1
wcwidth==0.2.13
wheel==0.45.1
whitenoise==6.8.2
